requirements.txt
tqdm==4.62.3
python-dotenv==0.20.0
peewee==3.14.8
pandas==1.3.5
xlrd==2.0.1
ipython-sql==0.4.0
matplotlib==3.5.1
seaborn==0.11.2
scikit-learn==1.0.2
Jinja2==3.1.1
statsmodels==0.13.2
plotly==5.7.0
dash==2.3.1